Remove external SDFatLib library dependency.
Currently, azip fails to compile on the stock Arduino v1.0 as it
requires the external SDFatLib library. However, since Arduino v1.0
includes a SD library based on the SDFatLib, supporting FAT
formatted SD cards, this is unecessary.

This commit migrates azip to the standard SD library, removing the
need to add the external library. This has been compiled,
uploaded and subjected to a basic test with the included Mini-Zork
with no obvious regressions.